refactor: extract grid pro edit select logic to a separate mixin
